+++ contents/config/config.qml
@@ -6,8 +6,13 @@ import org.kde.plasma.configuration
 
 ConfigModel {
     ConfigCategory {
-        name: i18n("General")
+        name: i18n("Appearance")
+        icon: "preferences-desktop-color"
+        source: "configAppearance.qml"
+    }
+    ConfigCategory {
+        name: i18n("Behavior")
         icon: "preferences-desktop"
-        source: "configGeneral.qml"
+        source: "configBehavior.qml"
     }
 }